added usedGuesses char list (read commit description + info)
So the only change made here was to add usedGuesses: char list because then we could cross-check if the char has been used before. I am not sure 100% is being checked right now. Anyway, as of now is a parameter that is not being used. Theoretically it could be deleted